Omit initial version of battery.py architecture
