+++ src/battery.py
@@ -4,11 +4,10 @@ from math import exp, log
 from src.loading_curve import LoadingCurve
 
 
-class Battery:
+class Battery():
     """Battery class"""
     def __init__(self, capacity, loading_curve, soc, efficiency=0.95, unloading_curve=None):
         """ Initializing the battery
-
         :param capacity: capacity of the battery
         :type capacity: int/float
         :param loading_curve: loading curve of the battery
@@ -44,47 +43,20 @@ class Battery:
         :rtype: dict
         """
 
-        # get interval in hours
-        total_time = timedelta.total_seconds() / 3600.0
-        # hours: available time for charging, initially complete timedelta
-        remaining_hours = total_time
-
         # get loading curve clamped to maximum value
         # adjust charging curve to reflect power that reaches the battery
         # after losses due to efficieny
         clamped = self.loading_curve.clamped(max_charging_power).scale(self.efficiency)
 
-        avg_power = 0
         old_soc = self.soc
 
-        # find current region in loading curve
-        _, idx_2 = clamped.get_linear_section(self.soc)
-        x2 = min(target_soc, clamped.points[idx_2][0])
-
-        energies = []
-
-        # compute average power for each linear section
-        # update SOC
-        # computes for whole time or until target is reached
-        # TODO: move loops to other function
-        while remaining_hours > self.EPS and target_soc - self.soc > self.EPS:  # self.soc < target:
-            while x2 - self.soc < self.EPS:  # self.soc >= x2:
-                # get next section
-                idx_2 += 1
-                x2 = min(target_soc, clamped.points[idx_2][0])
-
-            energy_delta, remaining_hours = self._adjust_soc(charging_curve=clamped,
-                                                             target_soc=x2,
-                                                             time_limit=remaining_hours)
-            # remember amount of energy loaded into battery
-            energies.append(energy_delta)
+        avg_power = self._adjust_soc(charging_curve=clamped,
+                                     target_soc=target_soc,
+                                     timedelta=timedelta)
 
         # get average power (energy over complete timedelta)
-        # supplied to the battery before loss due to efficieny
-        try:
-            avg_power = (sum(energies) / self.efficiency) / total_time
-        except ZeroDivisionError:
-            avg_power = 0
+        # supplied by the system to the connected device/vehicle after loss due to efficiency
+        avg_power /= self.efficiency
 
         return {'avg_power': avg_power, 'soc_delta': self.soc - old_soc}
 
@@ -109,46 +81,20 @@ class Battery:
         if max_power is None:
             max_power = self.unloading_curve.max_power
 
-        # get interval in hours
-        total_time = timedelta.total_seconds() / 3600.0
-        # hours: available time for charging, initially complete timedelta
-        remaining_hours = total_time
-
         # get loading curve clamped to maximum value
         # adjust loading curve by efficiency factor to reflect power
         # flowing out of the battery as opposed to power provided by the battery to user
         clamped = self.unloading_curve.clamped(max_power).scale(1/self.efficiency)
 
-        avg_power = 0
         old_soc = self.soc
 
-        # find initial linear section
-        idx_2, _ = clamped.get_linear_section(self.soc)
-        x2 = max(target_soc, clamped.points[idx_2][0])
-
-        energies = []
-
-        # compute average power for each linear section
-        # update SOC
-        # computes for whole time or until target is reached
-        while remaining_hours > self.EPS and self.soc - target_soc > self.EPS:  # self.soc > target:
-            while self.soc - x2 < self.EPS:  # self.soc <= x2:
-                # get next section
-                idx_2 -= 1
-                x2 = max(target_soc, clamped.points[idx_2][0])
-
-            energy_delta, remaining_hours = self._adjust_soc(charging_curve=clamped,
-                                                             target_soc=x2,
-                                                             time_limit=remaining_hours)
-            # remember amount of energy provided
-            energies.append(energy_delta)
+        avg_power = self._adjust_soc(charging_curve=clamped,
+                                     target_soc=target_soc,
+                                     timedelta=timedelta)
 
         # get average power (energy over complete timedelta)
         # supplied by the system to the connected device/vehicle after loss due to efficiency
-        try:
-            avg_power = (sum(energies) * self.efficiency) / total_time
-        except ZeroDivisionError:
-            avg_power = 0
+        avg_power *= self.efficiency
 
         return {'avg_power': avg_power, 'soc_delta':  old_soc - self.soc}
 
@@ -202,7 +148,7 @@ class Battery:
         self.soc = old_soc
         return power
 
-    def _adjust_soc(self, charging_curve, target_soc, time_limit):
+    def _adjust_soc(self, timedelta, charging_curve, target_soc):
         """ Helper function that loads or unloads battery to a given target SOC keeping track of
             the duration and stopping the process early if a time limit is reached.
 
@@ -216,54 +162,96 @@ class Battery:
         :type time_limit: _type_
         :return: _description_
         :rtype: _type_
-        """        # compute gradient and offset of linear equation
-
+        """
+        # compute gradient and offset of linear equation
         if abs(self.soc - target_soc) < self.EPS:
             # if target soc has already been reached, do not (dis)charge
-            energy_delta = 0
-            return energy_delta, time_limit
+            return 0
 
-        x1 = self.soc
-        x2 = target_soc
-        hours = time_limit
+        # get interval in hours
+        total_time = timedelta.total_seconds() / 3600.0
+        # hours: available time for charging, initially complete timedelta
+        remaining_hours = total_time
 
-        y1 = charging_curve.power_from_soc(x1)
-        y2 = charging_curve.power_from_soc(x2)
-        dx = x2 - x1
-        dy = y2 - y1
+        # for certain steps in the process below it matters whether
+        # the battery charges or discharges
+        discharge = target_soc < self.soc
 
-        m = dy / dx
-        n = y1 - m * x1
-        c = self.capacity
-
-        # find time to breakpoint
-        try:
-            if abs(m) < self.EPS:
-                # simple constant charging
-                t = (x2 - self.soc) * c / n
-            else:
-                # inverse of exponential function
-                t = log((x2 + n/m) / (self.soc + n/m)) * c/m
-        except (ValueError, ZeroDivisionError):
-            t = hours
-
-        # what is earlier, breakpoint or interval end?
-        # keep track of sign(t) as it encodes whether we charge or discharge
-        t = ((t >= 0) - (t < 0)) * min(abs(t), hours)
-
-        if abs(m) < self.EPS:
-            # simple case: charging with constant power, regardless of SOC
-            new_soc = self.soc + (n/c * t)
+        # find current region in loading curve
+        idx_1, idx_2 = charging_curve.get_linear_section(self.soc)
+        if discharge:
+            boundary_idx = idx_1
+            boundary_soc = max(target_soc, charging_curve.points[idx_1][0])
         else:
-            # charge power dependent on SOC
-            # inhomogenous differential equation -> exponential function
-            new_soc = -n/m + (n/m + self.soc) * exp(m/c * t)
+            boundary_idx = idx_2
+            boundary_soc = min(target_soc, charging_curve.points[idx_2][0])
 
-        energy_delta = abs(new_soc - self.soc) * c
-        self.soc = new_soc
-        hours -= abs(t)
+        # collect energy flowing in or out of battery
+        energies = []
 
-        return energy_delta, hours
+        sign = (-1)**discharge
+        # compute average power for each linear section
+        # update SOC
+        # computes for whole time or until target is reached
+
+        # self.soc < target if charging else self.soc > target
+        while remaining_hours > self.EPS and sign * (target_soc - self.soc) > self.EPS:
+            # self.soc >= boundary_soc if charging else self.soc <= boudary_soc
+            while sign * (boundary_soc - self.soc) < self.EPS:
+                # get next section
+                boundary_idx += sign
+                if discharge:
+                    boundary_soc = max(target_soc, charging_curve.points[boundary_idx][0])
+                else:
+                    boundary_soc = min(target_soc, charging_curve.points[boundary_idx][0])
+
+            x1 = self.soc
+            x2 = boundary_soc
+
+            y1 = charging_curve.power_from_soc(x1)
+            y2 = charging_curve.power_from_soc(x2)
+            dx = x2 - x1
+            dy = y2 - y1
+
+            m = dy / dx
+            n = y1 - m * x1
+            c = self.capacity
+
+            # find time to breakpoint
+            try:
+                if abs(m) < self.EPS:
+                    # simple constant charging
+                    t = (x2 - self.soc) * c / n
+                else:
+                    # inverse of exponential function
+                    t = log((x2 + n/m) / (self.soc + n/m)) * c/m
+            except (ValueError, ZeroDivisionError):
+                t = remaining_hours
+
+            # what is earlier, breakpoint or interval end?
+            # keep track of sign(t) as it encodes whether we charge or discharge
+            t = ((t >= 0) - (t < 0)) * min(abs(t), remaining_hours)
+
+            if abs(m) < self.EPS:
+                # simple case: charging with constant power, regardless of SOC
+                new_soc = self.soc + (n/c * t)
+            else:
+                # charge power dependent on SOC
+                # inhomogenous differential equation -> exponential function
+                new_soc = -n/m + (n/m + self.soc) * exp(m/c * t)
+
+            energy_delta = abs(new_soc - self.soc) * c
+            self.soc = new_soc
+            remaining_hours -= abs(t)
+            # remember amount of energy loaded into battery
+            energies.append(energy_delta)
+
+        try:
+            avg_power = sum(energies) / total_time
+        except ZeroDivisionError:
+            avg_power = 0
+
+        return avg_power
 
     def __str__(self):
         return 'Battery {}'.format({k: str(v) for k, v in vars(self).items()})
